+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 Here you can see the full list of changes between each GitHub Project PHP package release.
 
+## v1.0.0 - 2025-02-14
+
+### What's Changed
+
+- Handle comments for all field types in GitHub project V2.
+- Create an enable comment flag for the Status field.
+- Format color field values.
+
+**Full Changelog**: https://github.com/cslant/github-project-php/compare/v0.1.0...v1.0.0
+
 ## v0.1.0 - 2025-03-12
 
 - Experimental release